Small code changes to fix minor problems, and some changes in commenting of the code.

+++ Server/Forms/FrmRegValueEditWord.cs
@@ -87,7 +87,7 @@ namespace xServer.Forms
                         if (_value.Kind == RegistryValueKind.DWord)
                             return UInt32.MaxValue.ToString(returnType);
                         else
-                            return UInt32.MaxValue.ToString(returnType);
+                            return UInt64.MaxValue.ToString(returnType);
                     }
                 }
             }
